requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.1
django-crispy-forms==2.3
pillow==11.0.0
psycopg2-binary==2.9.9
sqlparse==0.5.1
uWSGI==2.0.26
